Changed a little the code, so it is better implemented into existing codebase(on android). Also changed the Qt private headers to their lower case, for some reason not always those headers are expose as the general Qt headers.
