+++ QtAddMobApp/QtAdMob/platform/android/gradle.properties
@@ -1,0 +1,9 @@
+## This file is automatically generated by QtCreator.
+#
+# This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+
+androidBuildToolsVersion=22.0.1
+androidCompileSdkVersion=22
+buildDir=.build
+qt5AndroidDir=/home/moisex/Qt5.5.0_android/5.5/android_armv7/src/android/java
