+++ QtAddMobApp/QtAdMob/platform/android/local.properties
@@ -1,0 +1,6 @@
+## This file is automatically generated by QtCreator.
+#
+# This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+
+sdk.dir=/home/moisex/android-sdks
